requirements.txt
SQLAlchemy==1.0.9
sqlalchemy-migrate==0.10.0
lxml==3.5.0
requests==2.8.1